# openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: info@librematch.org
    name: Libre:Match Team
    url: https://librematch.org
  description:
    "stable, reliable and feature-rich API for match information, leaderboards,\
    \ and tournament organisation for the Age of Empires multiplayer community"
  license:
    name: GNU AGPLv3
    url: https://www.gnu.org/licenses/agpl.txt
    x-last-modified: 1663947772699
  title: Libre:Match API
  version: 1.0.0-DRAFT
  x-logo:
    url: ""
servers:
  - description: Libre:Match Server v1
    url: https://api.librematch.org/v1
    x-last-modified: 1663947810574
security:
  - BasicAuth: []
  - X-Api-Key: []
paths:
...
  /application:
    get:
      operationId: get_information_collection_for_application
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
          x-last-modified: 1665056670316
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
      summary: Retrieve a list of available application information
      tags:
        - Internal
        - Administration
    summary:
      "General information and data about Libre:Match app, e.g. API updates,\
      \ possible upcoming maintenance and other news/stuff"
    x-last-modified: 1663948904071
  /application/statistics:
    get:
      operationId: get_stat_collection_for_application
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
          x-last-modified: 1665056670316
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
      summary: Retrieve a list of only application statistics
      tags:
        - Internal
        - Administration
    summary:
      "Statistics of Libre:Match, e.g. usage statistics, collection rate, overall\
      \ amount of users, etc."
    x-last-modified: 1663948946628
  /application/status:
    get:
      operationId: get_status_collection_for_application
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
          x-last-modified: 1665056670316
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
        "503":
          description: Service Unavailable
          x-last-modified: 1665057340292
      summary: Retrieve only application status
      tags:
        - Internal
        - Administration
        - Status Check
    summary: Status updates for health checks of our API
    x-last-modified: 1663948961011
...
components:
...
  schemas:
...
    TooManyRequestsError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 429
          minimum: 429
          title: code
          type: integer
      required:
        - code
        - message
      title: TooManyRequestsError
      type: object
      x-last-modified: 1665056674054
...
    NotAcceptableError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 406
          minimum: 406
          title: code
          type: integer
      required:
        - code
        - message
      title: NotAcceptableError
      type: object
      x-last-modified: 1665056674054
    NotFoundError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 404
          minimum: 404
          title: code
          type: integer
      required:
        - code
        - message
      title: NotFoundError
      type: object
      x-last-modified: 1665056674054
    ForbiddenError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 403
          minimum: 403
          title: code
          type: integer
      required:
        - code
        - message
      title: ForbiddenError
      type: object
      x-last-modified: 1665056674054
...
    UnauthorizedError:
      additionalProperties: false
      properties:
        message:
          maxLength: 32
          pattern: "^example-[0-9a-z]+$"
          title: Unauthorized access
          type: string
        code:
          maximum: 401
          minimum: 401
          title: code
          type: integer
      required:
        - code
        - message
      title: UnauthorizedError
      type: object
      x-last-modified: 1665056676293
...
  securitySchemes:
    BasicAuth:
      description: "Example: \n> Authorization: <Sample>"
      in: header
      name: BasicAuth
      type: apiKey
      x-last-modified: 1663947718456
    X-Api-Key:
      description: "Example: \n> X-Api-Key: ZGVtbzpwQDU1dzByZA=="
      in: header
      name: X-Api-Key
      type: apiKey
      x-last-modified: 1663947718456
    JWTAuth:
      bearerFormat: JWT
      description: "Example: \n> Authorization: Bearer <token>"
      scheme: bearer
      type: http
      x-last-modified: 1665071628782